src/requirements.txt
azure-functions
pymysql
